requirements.txt
cloudml-hypertune==0.1.0.dev6
google-pasta
google-cloud-storage
gsutil
huggingface-hub==0.11.1
nltk==3.7
numpy==1.23.4
pandas==1.5.2
pip==22.3.1
py-rouge==1.1
PyYAML==5.4.1
regex==2022.10.31
rouge-score==0.1.2
rsa==4.7.2
scikit-learn==0.22.1
scipy==1.9.3
tokenizers==0.13.0
torch==1.13.0
torchaudio==0.13.0
torchserve==0.6.1
torchvision==0.14.0
tqdm==4.64.1
transformers==4.25.1
Unidecode==1.3.6
urllib3==1.26.13
yacs==0.1.8
yarl==1.8.1
